requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.38.0
ddtrace>=2.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development dependencies
pytest>=7.4.3
pytest-mock>=3.12.0
pytest-asyncio>=0.21.1
ruff>=0.1.6
black>=23.11.0
mypy>=1.7.1
bandit[toml]>=1.7.5
pre-commit>=3.5.0

# Testing/mocking
responses>=0.24.1
